[SIMPLY-2369] Using data from authentication document to populate placeholder text for account detail entry
